requirements.txt
anyio==4.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.4
attrs==25.3.0
babel==2.18.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.14.3
bleach==6.1.0
blinker==1.8.2
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.0.3
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.4
jupyterlab==4.3.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.2.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.3
notebook_shim==0.2.4
overrides==7.7.0
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.35.1
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
Send2Trash==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.2.1
tomli==2.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
uri-template==1.3.0
urllib3==2.2.3
uWSGI==2.0.31
wcwidth==0.6.0
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.15
zipp==3.20.2
